+++ Sources/Moya/MoyaProvider.swift
@@ -53,10 +53,14 @@ open class MoyaProvider<Target: TargetType> {
 
     open internal(set) var inflightRequests: [Endpoint<Target>: [Moya.Completion]] = [:]
 
+    /// Propagated to Alamofire as callback queue. If nil - the Alamofire default (as of their API in 2017 - the main queue) will be used.
+    let callbackQueue: DispatchQueue?
+
     /// Initializes a provider.
     public init(endpointClosure: @escaping EndpointClosure = MoyaProvider.defaultEndpointMapping,
                 requestClosure: @escaping RequestClosure = MoyaProvider.defaultRequestMapping,
                 stubClosure: @escaping StubClosure = MoyaProvider.neverStub,
+                callbackQueue: DispatchQueue? = nil,
                 manager: Manager = MoyaProvider<Target>.defaultAlamofireManager(),
                 plugins: [PluginType] = [],
                 trackInflights: Bool = false) {
@@ -67,6 +71,7 @@ open class MoyaProvider<Target: TargetType> {
         self.manager = manager
         self.plugins = plugins
         self.trackInflights = trackInflights
+        self.callbackQueue = callbackQueue
     }
 
     /// Returns an `Endpoint` based on the token, method, and parameters by invoking the `endpointClosure`.
@@ -76,31 +81,36 @@ open class MoyaProvider<Target: TargetType> {
 
     /// Designated request-making method. Returns a `Cancellable` token to cancel the request later.
     @discardableResult
-    open func request(_ target: Target, completion: @escaping Moya.Completion) -> Cancellable {
-        return self.request(target, queue: nil, completion: completion)
-    }
+    open func request(_ target: Target,
+                      callbackQueue: DispatchQueue? = .none,
+                      progress: ProgressBlock? = .none,
+                      completion: @escaping Completion) -> Cancellable {
 
-    /// Designated request-making method with queue option. Returns a `Cancellable` token to cancel the request later.
-    @discardableResult
-    open func request(_ target: Target, queue: DispatchQueue?, progress: Moya.ProgressBlock? = nil, completion: @escaping Moya.Completion) -> Cancellable {
-        return requestNormal(target, queue: queue, progress: progress, completion: completion)
+        let callbackQueue = callbackQueue ?? self.callbackQueue
+        return requestNormal(target, callbackQueue: callbackQueue, progress: progress, completion: completion)
     }
 
     /// When overriding this method, take care to `notifyPluginsOfImpendingStub` and to perform the stub using the `createStubFunction` method.
     /// Note: this was previously in an extension, however it must be in the original class declaration to allow subclasses to override.
     @discardableResult
-    open func stubRequest(_ target: Target, request: URLRequest, completion: @escaping Moya.Completion, endpoint: Endpoint<Target>, stubBehavior: Moya.StubBehavior) -> CancellableToken {
+    open func stubRequest(_ target: Target, request: URLRequest, callbackQueue: DispatchQueue?, completion: @escaping Moya.Completion, endpoint: Endpoint<Target>, stubBehavior: Moya.StubBehavior) -> CancellableToken {
+        let callbackQueue = callbackQueue ?? self.callbackQueue
         let cancellableToken = CancellableToken { }
         notifyPluginsOfImpendingStub(for: request, target: target)
         let plugins = self.plugins
         let stub: () -> Void = createStubFunction(cancellableToken, forTarget: target, withCompletion: completion, endpoint: endpoint, plugins: plugins, request: request)
         switch stubBehavior {
         case .immediate:
-            stub()
+            switch callbackQueue {
+            case .none:
+                stub()
+            case .some(let callbackQueue):
+                callbackQueue.async(execute: stub)
+            }
         case .delayed(let delay):
             let killTimeOffset = Int64(CDouble(delay) * CDouble(NSEC_PER_SEC))
             let killTime = DispatchTime.now() + Double(killTimeOffset) / Double(NSEC_PER_SEC)
-            DispatchQueue.main.asyncAfter(deadline: killTime) {
+            (callbackQueue ?? DispatchQueue.main).asyncAfter(deadline: killTime) {
                 stub()
             }
         case .never: