+++ Tests/MoyaProviderIntegrationTests.swift
@@ -93,13 +93,13 @@ class MoyaProviderIntegrationTests: QuickSpec {
                         expect(manager.called) == true
                     }
 
-                    it("uses other background queue") {
+                    it("uses a background queue") {
                         var isMainThread: Bool?
-                        let queue = DispatchQueue(label: "background_queue", attributes: .concurrent)
+                        let callbackQueue = DispatchQueue(label: "background_queue", attributes: .concurrent)
                         let target: GitHub = .zen
 
                         waitUntil { done in
-                            provider.request(target, queue:queue) { _ in
+                            provider.request(target, callbackQueue: callbackQueue) { _ in
                                 isMainThread = Thread.isMainThread
                                 done()
                             }
@@ -108,7 +108,7 @@ class MoyaProviderIntegrationTests: QuickSpec {
                         expect(isMainThread) == false
                     }
 
-                    it("uses main queue") {
+                    it("uses the main queue") {
                         var isMainThread: Bool?
                         let target: GitHub = .zen