+++ Changelog.md
@@ -1,4 +1,6 @@
 # Next
+- **Breaking Change** Renamed all occurrences of `queue` to `callbackQueue`.
+- Add optional callback queue parameter to reactive providers.
 
 - Enabled the "Allow app extension API only" flag.
 - Added an optional `requestDataFormatter`in `NetworkLoggerPlugin` to allow the client to interact with the request data before logging it.